restructured repo, removed web3, added ethers
